Strauss {spatial}	R Documentation

Simulates Strauss Spatial Point Process

Description

Simulates Strauss spatial point process.

Usage


Strauss(n, c=0, r)

Arguments

n
number of points

c
parameter c in [0, 1]. c = 0 corresponds to complete inhibition at distances up to r.

r
inhibition distance

Details

Uses spatial birth-and-death process for 4n steps, or for 40n steps starting from a binomial pattern on the first call from an other function. Uses the region set by ppinit or ppregion.

Value

list of vectors of x and y coordinates

Side Effects

uses the random number generator

References

Ripley, B. D. (1981) Spatial Statistics. Wiley.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

Psim, SSI

Examples


towns <- ppinit("towns.dat")
par(pty="s")
plot(Kfn(towns, 10), type="b", xlab="distance", ylab="L(t)")
lines(Kaver(10, 25, Strauss(69,0.5,3.5)))

[Package spatial version 7.3-11 Index]