correlogram {spatial}	R Documentation

Compute Spatial Correlograms

Description

Compute spatial correlograms of spatial data or residuals.

Usage


correlogram(krig, nint, plotit = TRUE,  ...)

Arguments

krig
trend-surface or kriging object with columns x, y, and z

nint
number of bins used

plotit
logical for plotting

...
parameters for the plot

Details

Divides range of data into nint bins, and computes the covariance for pairs with separation in each bin, then divides by the variance. Returns results for bins with 6 or more pairs.

Value

x and y coordinates of the correlogram, and cnt, the number of pairs averaged per bin.

Side Effects

Plots the correlogram if plotit = TRUE.

References

Ripley, B. D. (1981) Spatial Statistics. Wiley.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

variogram

Examples


data(topo, package="MASS")
topo.kr <- surf.ls(2, topo)
correlogram(topo.kr, 25)
d <- seq(0, 7, 0.1)
lines(d, expcov(d, 0.7))

[Package spatial version 7.3-11 Index]