ppinit {spatial}	R Documentation

Read a Point Process Object from a File

Description

Read a file in standard format and create a point process object.

Usage


ppinit(file)

Arguments

file
string giving file name

Details

The file should contain

the number of points
a header (ignored)
xl xu yl yu scale
x y (repeated n times)

Value

class "pp" object with components x, y, xl, xu, yl, yu

Side Effects

Calls ppregion to set the domain.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

ppregion

Examples


towns <- ppinit("towns.dat")
par(pty="s")
plot(Kfn(towns, 10), type="b", xlab="distance", ylab="L(t)")

[Package spatial version 7.3-11 Index]