predict.trls {spatial}	R Documentation

Predict method for trend surface fits

Description

Predicted values based on trend surface model object

Usage


## S3 method for class 'trls'
predict(object, x, y, ...)

Arguments

object
Fitted trend surface model object returned by surf.ls

x
Vector of prediction location eastings (x coordinates)

y
Vector of prediction location northings (y coordinates)

...
further arguments passed to or from other methods.

Value

predict.trls produces a vector of predictions corresponding to the prediction locations. To display the output with image or contour, use trmat or convert the returned vector to matrix form.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

surf.ls, trmat

Examples


data(topo, package="MASS")
topo2 <- surf.ls(2, topo)
topo4 <- surf.ls(4, topo)
x <- c(1.78, 2.21)
y <- c(6.15, 6.15)
z2 <- predict(topo2, x, y)
z4 <- predict(topo4, x, y)
cat("2nd order predictions:", z2, "\n4th order predictions:", z4, "\n")

[Package spatial version 7.3-11 Index]