variogram {spatial}	R Documentation

Compute Spatial Variogram

Description

Compute spatial (semi-)variogram of spatial data or residuals.

Usage


variogram(krig, nint, plotit = TRUE, ...)

Arguments

krig
trend-surface or kriging object with columns x, y, and z

nint
number of bins used

plotit
logical for plotting

...
parameters for the plot

Details

Divides range of data into nint bins, and computes the average squared difference for pairs with separation in each bin. Returns results for bins with 6 or more pairs.

Value

x and y coordinates of the variogram and cnt, the number of pairs averaged per bin.

Side Effects

Plots the variogram if plotit = TRUE

References

Ripley, B. D. (1981) Spatial Statistics. Wiley.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

correlogram

Examples


data(topo, package="MASS")
topo.kr <- surf.ls(2, topo)
variogram(topo.kr, 25)

[Package spatial version 7.3-11 Index]